Removed display: flex for lost-column in flex-mode
When lost is in flexbox mode the attribute display:flex for lost-column is not necessary. It breaks it's child-elements layouts because they change to flex-items. When the parent element is a flex-box (e.g. display: flex) it's child-elements are automatically set to flex-elements. There's no need to set them on display:flex as well - only if it's own childs should be flex-elements.